+++ product-service/src/main/java/com/starshipshop/productservice/config/StarshipFeignConfiguration.java
@@ -1,29 +1,18 @@
 package com.starshipshop.productservice.config;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.starshipshop.productservice.client.StarshipFeignClient;
-import com.starshipshop.productservice.client.auth.BearerTokenRequestInterceptor;
 import feign.Client;
 import feign.Contract;
 import feign.Feign;
-import feign.auth.BasicAuthRequestInterceptor;
 import feign.codec.Decoder;
 import feign.codec.Encoder;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cloud.openfeign.security.OAuth2AccessTokenInterceptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @RequiredArgsConstructor
 public class StarshipFeignConfiguration {
-
-    @Value("${client.starship.username}")
-    private String username;
-    @Value("${client.starship.password}")
-    private String password;
 
     @Bean
     public StarshipFeignClient initFeignClient(final Decoder decoder, final Encoder encoder, final Client client, final Contract contract) {
@@ -32,7 +21,6 @@ public class StarshipFeignConfiguration {
                 .contract(contract)
                 .decoder(decoder)
                 .encoder(encoder)
-                .requestInterceptor(new BearerTokenRequestInterceptor())
                 .target(StarshipFeignClient.class, "starship-service");
     }
 }